move hardsubx context struct to mod.rs
